fix types time shift

+++ hw_asr/augmentations/wave_augmentations/time_shift.py
@@ -1,5 +1,4 @@
 import numpy as np
-import torch_audiomentations
 from torch import Tensor
 import torch
 
@@ -16,9 +15,9 @@ class TimeShift(AugmentationBase):
         start_ = int(np.random.uniform(self.start, self.stop))
         data = data.squeeze(0)
         if start_ >= 0:
-            wav_time_shift = np.r_[data[start_:], np.random.uniform(-0.001, 0.001, start_)]
+            wav_time_shift = np.r_[data[start_:], np.random.uniform(-0.001, 0.001, start_).astype(np.double)]
         else:
-            wav_time_shift = np.r_[np.random.uniform(-0.001, 0.001, -start_), data[:start_]]
+            wav_time_shift = np.r_[np.random.uniform(-0.001, 0.001, -start_).astype(np.double), data[:start_]]
         if not torch.is_tensor(wav_time_shift):
             wav_time_shift = torch.from_numpy(wav_time_shift)
         return wav_time_shift.unsqueeze(0)
